Move and include BeanMethodsNotPublic in best practices
