+++ src/main/resources/META-INF/rewrite/spring-boot-20.yml
@@ -183,3 +183,4 @@ recipeList:
   - org.openrewrite.java.spring.NoAutowiredOnConstructor
   - org.openrewrite.java.spring.boot2.RestTemplateBuilderRequestFactory
   - org.openrewrite.java.spring.boot2.ReplaceDeprecatedEnvironmentTestUtils
+  - org.openrewrite.java.spring.framework.BeanMethodsNotPublic
