+++ src/main/resources/META-INF/rewrite/spring-boot-30.yml
@@ -71,7 +71,7 @@ recipeList:
   - org.openrewrite.java.spring.boot3.MigrateDropWizardDependencies
   - org.openrewrite.java.spring.boot3.RemoveSolrAutoConfigurationExclude
   - org.openrewrite.java.spring.boot3.MigrateWebMvcTagsToObservationConvention
-  - org.openrewrite.java.spring.boot3.BeanMethodReturnNull
+  - org.openrewrite.java.spring.framework.BeanMethodReturnNull
   - org.openrewrite.java.spring.batch.SpringBatch4To5Migration
   - org.openrewrite.java.spring.framework.UpgradeSpringFramework_6_0
   - org.openrewrite.java.spring.kafka.UpgradeSpringKafka_3_0
